+++ src/app/controllers/webhook/message.ts
@@ -1,8 +1,6 @@
 /**
  * LINE webhook messageコントローラー
- * @namespace app.controllers.webhook.message
  */
-
 import * as sskts from '@motionpicture/sskts-domain';
 import * as createDebug from 'debug';
 import * as moment from 'moment';
@@ -15,7 +13,6 @@ const debug = createDebug('sskts-line-assistant:controller:webhook:message');
 
 /**
  * 使い方を送信する
- * @export
  */
 export async function pushHowToUse(userId: string) {
     // tslint:disable-next-line:no-multiline-string
@@ -81,7 +78,6 @@ export async function askTransactionInquiryKey(user: User) {
 
 /**
  * 予約番号or電話番号のボタンを送信する
- * @export
  */
 export async function pushButtonsReserveNumOrTel(userId: string, message: string) {
     debug(userId, message);
@@ -130,7 +126,6 @@ export async function pushButtonsReserveNumOrTel(userId: string, message: string
 
 /**
  * 日付選択を求める
- * @export
  */
 export async function askFromWhenAndToWhen(userId: string) {
     // await LINE.pushMessage(userId, '期間をYYYYMMDD-YYYYMMDD形式で教えてください。');
@@ -168,7 +163,6 @@ export async function askFromWhenAndToWhen(userId: string) {
 
 /**
  * 取引CSVダウンロードURIを発行する
- * @export
  */
 export async function publishURI4transactionsCSV(userId: string, dateFrom: string, dateThrough: string) {
     await LINE.pushMessage(userId, `${dateFrom}-${dateThrough}の取引を検索しています...`);